+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add commands for attaching, detaching, and listing policies attached to a user in managed object storage service.
+- Add command for listing and deleting file-storage instances.
 
 ## [3.29.0] - 2026-02-06